added config volume

+++ root/usr/local/bin/run-container.sh
@@ -1,6 +1,17 @@
 #!/bin/bash
 
-rm /var/run/dbus/pid #incase shutdown abruptly
+if [ -n "$PLUGINS" ]; then
+  _PLUGINS=$(echo $PLUGINS | tr ";" "\n")
+  for PLUGIN in $_PLUGINS
+  do
+      "Installing plugin: ${PLUGIN}"
+      npm install -g ${PLUGIN}
+  done
+fi
+
+if [ -f /var/run/dbus/pid ]; then
+  rm /var/run/dbus/pid #incase shutdown abruptly
+fi
 
 sed -i "s/rlimit-nproc=3/#rlimit-nproc=3/" /etc/avahi/avahi-daemon.conf
